+++ src/themes/_generator/utilities/generateColors.ts
@@ -66,8 +66,10 @@ const generateColorValues = (args: { key: string; hex: string; hexDark?: string 
 	const bgHex = hex;
 	const bgHexDark = okhslToHex(okhslDark);
 
-	const bgFadedHsl = { ...okhsl, l: 0.94 + 0.045 * hueLightness };
-	const bgFadedHslDark = { ...okhslDark, l: 0.16, s: okhslDark.s / 2 };
+	const bgFadedHsl =
+		key === "neutral" ? { ...okhsl, l: 0.96 } : { ...okhsl, l: 0.94 + 0.045 * hueLightness };
+	const bgFadedHslDark =
+		key === "neutral" ? { ...okhslDark, l: 0.15 } : { ...okhslDark, l: 0.16, s: okhslDark.s / 2 };
 	const bgFadedHex = okhslToHex(bgFadedHsl);
 	const bgFadedHexDark = okhslToHex(bgFadedHslDark);